Add a new configuration options for HTTP timeouts
These configurations set the read_/open_/ssl_timeout attributes on the Net::HTTP instance used to make get requests.

The changes to the Client class are necessary to expose the http client before we issue the request.

Re: testing, there are no additions to the client_spec.rb because WebMock does not provide adequate facility to test particular values of the timeout settings.